@@ -4,5 +4,5 @@ var babel = require('gulp-babel')
 gulp.task('default', function () {
     return gulp.src('./src/**')
         .pipe(babel())
-        .pipe(gulp.dest('./lib'));
+        .pipe(gulp.dest('./lib'))
 });
