@@ -1,7 +1,8 @@
+
 module.exports = {
     entry: './index.jsx',
     output: {
-        publicPath: 'http://localhost:9090/assets'
+        publicPath: 'http://localhost:8091/assets'
     },
     module: {
         loaders: require('./loaders.config')
@@ -10,6 +11,7 @@ module.exports = {
         'react': 'React'
     },
     resolve: {
+        // Allow to omit extensions when requiring these files
         extensions: ['', '.js', '.jsx']
     }
 }
